Commit LCS Anagram
since this is just finding the length of the LCS rather than finding the LCS through Dynamic programming - we can just use this. 
The normal method would be to sort both the strings and find the common ones by merge sort - O(worst) = mergesort of both strings+ iterating through both the strings once. 
- this given solution is the best  O(n+m) solution